requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==1.31.0
google-auth==1.33.0
google-cloud-core==1.7.1
google-cloud-datastore==2.1.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.1
idna==3.2
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
libcst==0.3.19
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
neovim==0.3.1
packaging==21.0
parso==0.8.2
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pynvim==0.4.3
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.21
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
Werkzeug==2.0.1
